refactor: Organize project
