--- BarcodeScannerSwiftUI/BarcodeScannerView.swift
@@ -1,65 +1,0 @@
-//
-//  ContentView.swift
-//  BarcodeScannerSwiftUI
-//
-//  Created by Oscar Cristaldo on 2022-07-13.
-//
-
-import SwiftUI
-
-struct AlertItem: Identifiable {
-    let id = UUID()
-    let title: String
-    let message: String
-    let dismissButton: Alert.Button
-}
-
-struct AlertContext {
-    static let invalidDeviceInput = AlertItem(title: "Invalid device input",
-                                              message: "Something is wrong with the camera. We are unable to capture the input.",
-                                              dismissButton: .default(Text("Ok")))
-    
-    static let invalidScannedType = AlertItem(title: "Invalid scanned type",
-                                              message: "The value scanned is not valid. This app scans EAN-8 and EAN-13.",
-                                              dismissButton: .default(Text("Ok")))
-}
-
-struct BarcodeScannerView: View {
-    
-    @State private var scannedCode = ""
-    @State private var alertItem: AlertItem?
-    
-    var body: some View {
-        NavigationView {
-            VStack {
-                ScannerView(scannedCode: $scannedCode, alertItem: $alertItem)
-                    .frame(maxWidth: .infinity, maxHeight: 300)
-                
-                Spacer()
-                    .frame(height: 60)
-                
-                Label("Scanned Barcode:", systemImage: "barcode.viewfinder")
-                    .font(.title)
-                
-                Text(scannedCode.isEmpty ? "Not Yet Scanned" : scannedCode)
-                    .bold()
-                    .font(.largeTitle)
-                    .foregroundColor(scannedCode.isEmpty ? .red : .green)
-                    .padding()
-            }
-            .navigationTitle("Barcode Scanner")
-            .alert(item: $alertItem) { alertItem in
-                Alert(title: Text(alertItem.title),
-                      message: Text(alertItem.message),
-                      dismissButton: alertItem.dismissButton)
-            }
-            
-        }
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        BarcodeScannerView()
-    }
-}
