+++ BarcodeScannerSwiftUI/Screens/BarcodeScanner/BarcodeScannerViewModel.swift
@@ -1,0 +1,21 @@
+//
+//  BarcodeScannerViewModel.swift
+//  BarcodeScannerSwiftUI
+//
+//  Created by Oscar Cristaldo on 2022-07-13.
+//
+
+import SwiftUI
+
+final class BarcodeScannerViewModel: ObservableObject {
+    @Published var scannedCode = ""
+    @Published var alertItem: AlertItem?
+    
+    var statusText: String {
+        scannedCode.isEmpty ? "Not Yet Scanned" : scannedCode
+    }
+    
+    var statusColor: Color {
+        scannedCode.isEmpty ? .red : .green
+    }
+}
