+++ BarcodeScannerSwiftUI/Utils/Alert.swift
@@ -1,0 +1,25 @@
+//
+//  Alert.swift
+//  BarcodeScannerSwiftUI
+//
+//  Created by Oscar Cristaldo on 2022-07-13.
+//
+
+import SwiftUI
+
+struct AlertItem: Identifiable {
+    let id = UUID()
+    let title: String
+    let message: String
+    let dismissButton: Alert.Button
+}
+
+struct AlertContext {
+    static let invalidDeviceInput = AlertItem(title: "Invalid device input",
+                                              message: "Something is wrong with the camera. We are unable to capture the input.",
+                                              dismissButton: .default(Text("Ok")))
+    
+    static let invalidScannedType = AlertItem(title: "Invalid scanned type",
+                                              message: "The value scanned is not valid. This app scans EAN-8 and EAN-13.",
+                                              dismissButton: .default(Text("Ok")))
+}
