+++ BarcodeScannerSwiftUI/Views/UIKit Components/ScannerView.swift
